move PostType classes to dedicated namespace
