+++ src/Post/Model.php
@@ -7,6 +7,7 @@ use WP_Post;
 use WP_Query;
 use Illuminate\Support\Collection;
 use Silk\Query\Builder;
+use Silk\PostType\PostType;
 use Silk\Database\ActiveRecord;
 use Silk\Post\Exception\PostNotFoundException;
 use Silk\Post\Exception\ModelPostTypeMismatchException;
