+++ tests/unit/PostType/PostTypeBuilderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use Silk\Post\PostTypeBuilder;
+use Silk\PostType\Builder;
 
 class PostTypeBuilderTest extends WP_UnitTestCase
 {
@@ -11,7 +11,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
      */
     function it_can_be_constructed_with_a_slug()
     {
-        new PostTypeBuilder('some-type');
+        new Builder('some-type');
     }
 
     /**
@@ -19,7 +19,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
     */
     function it_has_a_named_constructor_for_creating_a_new_instance()
     {
-        $this->assertInstanceOf(PostTypeBuilder::class, PostTypeBuilder::make('new-type'));
+        $this->assertInstanceOf(Builder::class, Builder::make('new-type'));
     }
 
     /**
@@ -29,7 +29,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
     {
         $this->assertPostTypeNotExists('some-post-type');
 
-        $object = PostTypeBuilder::make('some-post-type')->register()->object();
+        $object = Builder::make('some-post-type')->register()->object();
 
         $this->assertSame($object, get_post_type_object('some-post-type'));
 
@@ -42,7 +42,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
     */
     function it_blows_up_if_the_post_type_slug_is_too_long()
     {
-        PostTypeBuilder::make('twenty-character-limit');
+        Builder::make('twenty-character-limit');
     }
 
     /**
@@ -51,7 +51,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
     */
     function it_blows_up_if_the_post_type_slug_is_too_short()
     {
-        PostTypeBuilder::make('');
+        Builder::make('');
     }
 
     /**
@@ -59,12 +59,12 @@ class PostTypeBuilderTest extends WP_UnitTestCase
      */
     function it_accepts_an_array_or_parameters_for_supported_features()
     {
-        PostTypeBuilder::make('bread')->supports(['flour', 'water'])->register();
+        Builder::make('bread')->supports(['flour', 'water'])->register();
 
         $this->assertTrue(post_type_supports('bread', 'flour'));
         $this->assertTrue(post_type_supports('bread', 'water'));
 
-        PostTypeBuilder::make('butter')->supports('bread', 'spreading')->register();
+        Builder::make('butter')->supports('bread', 'spreading')->register();
 
         $this->assertTrue(post_type_supports('butter', 'bread'));
         $this->assertTrue(post_type_supports('butter', 'spreading'));
@@ -75,7 +75,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
      */
     function it_can_get_and_set_arbitrary_values_for_the_registration_arguments()
     {
-        $type = PostTypeBuilder::make('stuff')
+        $type = Builder::make('stuff')
             ->set('mood', 'happy')
             ->register();
 
@@ -89,10 +89,10 @@ class PostTypeBuilderTest extends WP_UnitTestCase
      */
     function it_has_dedicated_methods_for_public_visibility()
     {
-        $public = PostTypeBuilder::make('a-public-type')->open();
+        $public = Builder::make('a-public-type')->open();
         $this->assertTrue($public->get('public'));
 
-        $private = PostTypeBuilder::make('a-private-type')->closed();
+        $private = Builder::make('a-private-type')->closed();
         $this->assertFalse($private->get('public'));
     }
 
@@ -101,10 +101,10 @@ class PostTypeBuilderTest extends WP_UnitTestCase
      */
     function it_has_dedicated_methods_for_user_interface()
     {
-        $ui = PostTypeBuilder::make('ui-having')->withUI();
+        $ui = Builder::make('ui-having')->withUI();
         $this->assertTrue($ui->get('show_ui'));
 
-        $no_ui = PostTypeBuilder::make('no-ui')->noUI();
+        $no_ui = Builder::make('no-ui')->noUI();
         $this->assertFalse($no_ui->get('show_ui'));
     }
 
@@ -113,7 +113,7 @@ class PostTypeBuilderTest extends WP_UnitTestCase
      */
     function it_has_methods_for_setting_the_labels()
     {
-        $book = PostTypeBuilder::make('book')
+        $book = Builder::make('book')
             // override a default value
             ->setLabel('archives', 'All the Bookz')
             // override a default with a new placeholder
